tercer commit

+++ practica-lunes2/src/main/java/com/example/practicalunes2/model/ReservaDTO.java
@@ -4,11 +4,11 @@ import java.time.LocalDate;
 
 public class ReservaDTO {
 	
-	private Long id;
+	private Integer id;
 
-	private Long libroId;
+	private Integer libroId;
 
-	private Long usuarioId;
+	private Integer usuarioId;
 
 	private LocalDate fechaReserva;
